# Conflicts:
#	kb-api/src/main/java/com/sb/solutions/api/loan/service/CustomerLoanService.java
#	kb-api/src/main/java/com/sb/solutions/api/loan/service/CustomerLoanServiceImpl.java

+++ kb-api/src/main/java/com/sb/solutions/api/proposal/entity/Proposal.java
@@ -54,7 +54,7 @@ public class Proposal extends BaseEntity<Long> {
     @NotAudited
     private String cashMarginMethod;
     @NotAudited
-    private Double enhanceLimitAmount;
+    private BigDecimal enhanceLimitAmount;
 
 
     @Transient
